// src/Model/User.ts
export class User {
         userId?: any;
         mail: string;
         passWord: string;
         userName: string;
         photoUser?: any;
         dateDeNaissance?: any;
         ville: string;
         codePostal: number;
         adresse: string;
         userSex: string;
         userLogement?: any;
         phone?:any;
       }