// src/Model/Logement.ts
import { User } from "./User";

export class Logement {
         logementid?: any;
         surfaceDELogement: number;
         descriptionDeLogement: string;
         photoLogement_1: any;
         photoLogement_2?: any;
         photoLogement_3?: any;
         ville: string;
         codePostal: number;
         adresse: string;
         proprietaire_id: any;
         dateDeAjoute: any;
         disponiblite?: boolean;
       }